Modified intent from uint8 to uint16

+++ server/intent.go
@@ -3,4 +3,4 @@ package server
 const INTENT_SERVER_SEND uint16 = 0x55
 const INTENT_SERVER_RECIEVE uint16 = 0x59
 
-const INTENT_STREAM_SIZE uint8 = 2
+const INTENT_STREAM_SIZE uint16 = 2
